requirements.txt
argon2-cffi==19.2.0
asgiref==3.2.7
asn1crypto==1.3.0
astroid==2.3.3
bcrypt==3.1.7
bokeh==2.0.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
defusedxml==0.6.0
Django==3.0.3
django-allauth==0.41.0
django-filter==2.2.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
Faker==2.0.4
idna==2.9
isort==4.3.21
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Markdown==3.2.1
MarkupSafe==1.1.1
mccabe==0.6.1
mkl-fft==1.0.15
mkl-service==2.3.0
numpy==1.18.1
oauthlib==3.1.0
olefile==0.46
packaging==20.3
Pillow==7.0.0
pycparser==2.20
pylint==2.5.0
pyOpenSSL==19.1.0
pyparsing==2.4.6
PySocks==1.7.1
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.2
tornado==6.0.4
typing-extensions==3.7.4.1
urllib3==1.25.8
wrapt==1.12.1
